src/requirements.txt
supabase==2.27.0
httpx==0.28.1
postgrest==2.27.0
realtime==2.27.0
storage3==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
yarl==1.22.0
httpcore==1.0.9
h11==0.16.0
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0